+++ tests/__init__.py
@@ -1,1 +1,5 @@
 from pyLinguist.utils import Logger
+from pyLinguist import (
+    Translator, Dictionary, Predictor, Speller, YaTranslateException,
+    YaBaseAPIHandler
+)
